Accessing data from an Azure sql database
I created a user interface to show data from a table stored in a database in azure
This interface contains several filters to show differents products and many options as showing products wit a different
language and availability of products.
This projects has an issue so when a product is double clicked it shows a new dialog where we can see more information about
the product and has an option for updating the product as the user wants, but this feature not working due to an
wrong implementation for the products information. This issue can be solved if we get the id of the products so we can
update the unique product.